feat: make jrpc rust client
