+++ applications/tari_validator_node/src/config.rs
@@ -144,7 +144,7 @@ impl Default for ValidatorNodeConfig {
             data_dir: PathBuf::from("data/validator_node"),
             p2p,
             grpc_address: Some("/ip4/127.0.0.1/tcp/18144".parse().unwrap()),
-            json_rpc_address: Some("127.0.0.1:18145".parse().unwrap()),
+            json_rpc_address: Some("127.0.0.1:18200".parse().unwrap()),
             http_ui_address: Some("127.0.0.1:5000".parse().unwrap()),
 
             // TODO: Deprecated - need to update tari_common to remove these
