+++ applications/tari_validator_node_cli/Cargo.toml
@@ -9,9 +9,10 @@ edition = "2018"
 
 [dependencies]
 tari_utilities = { git = "https://github.com/tari-project/tari_utilities.git", tag = "v0.4.7" }
-tari_dan_engine = { path = "../../dan_layer/engine"}
-tari_dan_common_types = { path = "../../dan_layer/common_types"}
-tari_engine_types = {path = "../../dan_layer/engine_types"}
+tari_dan_engine = { path = "../../dan_layer/engine" }
+tari_dan_common_types = { path = "../../dan_layer/common_types" }
+tari_engine_types = { path = "../../dan_layer/engine_types" }
+tari_validator_node_client = { path = "../../clients/validator_node_client" }
 
 anyhow = "1.0.65"
 clap = { version = "3.2.22", features = ["derive"] }
