--- applications/tari_validator_node_grpc/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "tari_validator_node_grpc"
-authors = ["The Tari Development Community"]
-repository = "https://github.com/tari-project/tari"
-license = "BSD-3-Clause"
-version = "0.35.0"
-edition = "2018"
-
-[dependencies]
-tari_common_types = {git = "https://github.com/tari-project/tari.git", branch = "feature-dan", package = "tari_common_types"}
-tari_crypto = { git = "https://github.com/tari-project/tari-crypto.git", tag = "v0.15.5" }
-tari_utilities = { git = "https://github.com/tari-project/tari_utilities.git", tag = "v0.4.5" }
-tari_dan_engine = { path = "../../dan_layer/engine"}
-tari_template_lib = { path = "../../dan_layer/template_lib"}
-
-chrono = { version = "0.4.19", default-features = false }
-prost = "0.9"
-prost-types = "0.9"
-num-traits = "0.2.15"
-tonic = "0.6.2"
-borsh = { version ="0.9.3", default-features = false }
-
-[build-dependencies]
-tonic-build = "0.6.2"
-
