+++ clients/validator_node_client/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tari_validator_node_client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tari_dan_common_types = { path = "../../dan_layer/common_types"}
+
+anyhow = "1.0.65"
+reqwest = { version = "0.11.11", features = ["json"] }
+serde = "1.0.144"
+serde_json = "1.0.85"
