--- clients/validator_node_grpc_client/README.md
@@ -1,19 +1,0 @@
-# Tari Validator Node GRPC client
-
-Async GRPC client library for the Tari Validator Node.
-
-## Usage
-
-```javascript
-const { Client } = require("@tari/validator-node-grpc-client");
-const client = Client.connect("127.0.0.1:18144");
-```
-
-## Development
-
-- `npm install`
-- hack hack hack
-
-## Tests
-
-- `npm test`
