--- clients/validator_node_grpc_client/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@tari/validator-node-grpc-client",
-  "version": "0.0.1",
-  "description": "GRPC client for Tari Base Node",
-  "main": "./src/index.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "author": "Tari development community",
-  "dependencies": {
-    "@grpc/grpc-js": "^1.2.3",
-    "@grpc/proto-loader": "^0.5.5",
-    "grpc-promise": "^1.4.0"
-  },
-  "devDependencies": {
-    "jest": "^27.1.1"
-  }
-}
